Revert "Revert "transportation/status screen improvements & image population""
