+++ android/app/google-services.json
@@ -1,0 +1,55 @@
+{
+  "project_info": {
+    "project_number": "956828958362",
+    "firebase_url": "https://books-and-breakfast-default-rtdb.firebaseio.com",
+    "project_id": "books-and-breakfast",
+    "storage_bucket": "books-and-breakfast.appspot.com"
+  },
+  "client": [
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:956828958362:android:c9f7132e76468cf7f96923",
+        "android_client_info": {
+          "package_name": "com.booksandbreakfast.booksandbreakfast"
+        }
+      },
+      "oauth_client": [
+        {
+          "client_id": "956828958362-c8hhnhdaq8eav386975ev1lftrhbmj18.apps.googleusercontent.com",
+          "client_type": 1,
+          "android_info": {
+            "package_name": "com.booksandbreakfast.booksandbreakfast",
+            "certificate_hash": "bcf4092e1972a5f6dbf98c4d91b5756e89a0eb2e"
+          }
+        },
+        {
+          "client_id": "956828958362-rb5qbi1g0rfm1l725no0rkn2mlmjonou.apps.googleusercontent.com",
+          "client_type": 3
+        }
+      ],
+      "api_key": [
+        {
+          "current_key": "AIzaSyBvUNf25valiVMKVs3GromLOlWoWy1kBtc"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": [
+            {
+              "client_id": "956828958362-fkvftsjf2vrjqh1ghs1cpiroipa5i2d0.apps.googleusercontent.com",
+              "client_type": 3
+            },
+            {
+              "client_id": "956828958362-inregq9gk1qhe2h8ar7pe8kra8opr71a.apps.googleusercontent.com",
+              "client_type": 2,
+              "ios_info": {
+                "bundle_id": "com.booksandbreakfast.booksandbreakfast"
+              }
+            }
+          ]
+        }
+      }
+    }
+  ],
+  "configuration_version": "1"
+}
