+++ app/(tabs)/_layout.tsx
@@ -30,18 +30,9 @@ export default function TabLayout() {
         }}
       />
       <Tabs.Screen
-        name="signup"
+        name="transportation"
         options={{
-          title: 'Sign up',
-          tabBarIcon: ({ color }) => (
-            <Ionicons name="calendar-clear-sharp" size={24} color={color} />
-          ),
-        }}
-      />
-      <Tabs.Screen
-        name="rideshare"
-        options={{
-          title: 'Rideshare',
+          title: 'Transportation',
           tabBarIcon: ({ color }) => <Ionicons name="car" size={28} color={color} />,
         }}
       />
