+++ app/components/MissionScreen.tsx
@@ -1,10 +1,11 @@
 import React, { useEffect, useState } from 'react';
-import { Dimensions, ScrollView, StyleSheet, Text, View } from 'react-native';
+import { Dimensions, ScrollView, StyleSheet, Text, View, Image } from 'react-native';
 import Carousel from 'react-native-reanimated-carousel';
 
 import { Entry, getMissionEntries } from '../firebase/util';
 
 const { width: screenWidth } = Dimensions.get('window');
+const screenHeight = Dimensions.get('window').height;
 
 const MissionScreen: React.FC = () => {
   const [entries, setEntries] = useState<Entry[]>([]);
@@ -23,20 +24,31 @@ const MissionScreen: React.FC = () => {
   }, []);
 
   const renderItem = ({ item }: { item: Entry }) => (
-    <ScrollView style={styles.item}>
-      <Text style={styles.title}>{item.title}</Text>
-      {item.subtitle && <Text style={styles.subtitle}>{item.subtitle}</Text>}
-      {item.missionCenter && <Text style={styles.missionCenter}>{item.missionCenter}</Text>}
-      <Text style={styles.body}>{item.body}</Text>
-      {item.subtitle1 && <Text style={styles.subtitle}>{item.subtitle1}</Text>}
-      {item.body1 && <Text style={styles.body}>{item.body1}</Text>}
-      {item.subtitle2 && <Text style={styles.subtitle}>{item.subtitle2}</Text>}
-      {item.body2 && <Text style={styles.body}>{item.body2}</Text>}
-      {item.subtitle3 && <Text style={styles.subtitle}>{item.subtitle3}</Text>}
-      {item.body3 && <Text style={styles.body}>{item.body3}</Text>}
-      {item.subtitle4 && <Text style={styles.subtitle}>{item.subtitle4}</Text>}
-      {item.body4 && <Text style={styles.body}>{item.body4}</Text>}
-    </ScrollView>
+    <View>
+      <ScrollView contentContainerStyle={styles.item}>
+        <Text style={styles.title}>{item.title}</Text>
+        {item.subtitle && <Text style={styles.subtitle}>{item.subtitle}</Text>}
+        {item.missionCenter && <Text style={styles.missionCenter}>{item.missionCenter}</Text>}
+        {item.missionCenter && (
+          <Image source={require('../../assets/mission-stmt-1.jpg')} style={styles.image} />
+        )}
+        {item.missionCenter && (
+          <Image source={require('../../assets/mission-stmt-2.jpeg')} style={styles.image} />
+        )}
+        <Text style={styles.body}>{item.body}</Text>
+        {item.subtitle1 && <Text style={styles.subtitle}>{item.subtitle1}</Text>}
+        {item.body1 && <Text style={styles.body}>{item.body1}</Text>}
+        {item.subtitle2 && <Text style={styles.subtitle}>{item.subtitle2}</Text>}
+        {item.body2 && <Text style={styles.body}>{item.body2}</Text>}
+        {item.subtitle3 && <Text style={styles.subtitle}>{item.subtitle3}</Text>}
+        {item.body3 && <Text style={styles.body}>{item.body3}</Text>}
+        {item.body3 && !item.subtitle4 && (
+          <Image source={require('../../assets/site-director.jpg')} style={styles.image} />
+        )}
+        {item.subtitle4 && <Text style={styles.subtitle}>{item.subtitle4}</Text>}
+        {item.body4 && <Text style={styles.body}>{item.body4}</Text>}
+      </ScrollView>
+    </View>
   );
 
   return (
@@ -47,6 +59,7 @@ const MissionScreen: React.FC = () => {
         width={screenWidth}
         onSnapToItem={(index) => setActiveSlide(index)}
         loop={false}
+        pagingEnabled
       />
       <View style={styles.paginationContainer}>
         {entries.map((_, index) => (
@@ -69,18 +82,21 @@ const styles = StyleSheet.create({
     borderRadius: 5,
     padding: 30,
     marginBottom: 20, // Added marginBottom to separate items if needed
+    flexGrow: 1,
+    minHeight: screenHeight,
   },
   subtitle: {
     fontSize: 15,
     fontWeight: 'bold',
     textDecorationLine: 'underline',
-    marginBottom: 5, // Added marginBottom for spacing
+    marginBottom: 10, // Added marginBottom for spacing
   },
   missionCenter: {
     fontSize: 15,
     color: '#333',
     textAlign: 'center', // Correct property to center text
     marginBottom: 5, // Added marginBottom for spacing
+    lineHeight: 20,
   },
   title: {
     fontSize: 26,
@@ -93,6 +109,7 @@ const styles = StyleSheet.create({
     fontSize: 15,
     color: '#333',
     marginBottom: 10, // Added marginBottom for spacing
+    lineHeight: 20,
   },
   paginationContainer: {
     flexDirection: 'row',
@@ -106,6 +123,12 @@ const styles = StyleSheet.create({
     borderRadius: 5,
     marginHorizontal: 3,
     backgroundColor: 'rgba(95, 87, 79, 0.92)',
+  },
+  image: {
+    width: 325,
+    height: 225,
+    marginTop: 20,
+    marginLeft: 12,
   },
 });
