+++ app/components/SchoolTransportDetails.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import { Dimensions, Pressable, ScrollView, StyleSheet, Text, View } from 'react-native';
+import { Image, Dimensions, Pressable, ScrollView, StyleSheet, Text, View } from 'react-native';
 import { TouchableOpacity } from 'react-native-gesture-handler';
 import { showLocation } from 'react-native-map-link';
 import FaIcon from 'react-native-vector-icons/FontAwesome';
@@ -11,6 +11,7 @@ import {
   updateSchoolDirections,
 } from '../firebase/util';
 import EditText from './EditText';
+import { ScreenWidth } from 'react-native-elements/dist/helpers';
 
 const screenHeight = Dimensions.get('window').height;
 
@@ -126,13 +127,16 @@ export const SchoolTransportDetails: React.FC<SchoolTransportDetailsProps> = ({
       <ScrollView contentContainerStyle={styles.scrollViewContentContainer}>
         {directionsInfo && (
           <>
-            <ArticleHeader
-              schoolName={directionsInfo.schoolName}
-              location={directionsInfo.address}
-              latitude={directionsInfo.geoLat}
-              longitude={directionsInfo.geoLong}
-            />
-            <View style={styles.section}>
+            <Image source={require('../../assets/bus-header.jpg')} style={styles.articleHeader}/>
+            <View style={styles.overlay}>
+              <ArticleHeader
+                schoolName={directionsInfo.schoolName}
+                location={directionsInfo.address}
+                latitude={directionsInfo.geoLat}
+                longitude={directionsInfo.geoLong}
+              />
+            </View>
+            <View style={styles.firstSection}>
               <Text style={styles.header}>Directions</Text>
               <EditText
                 value={directionsInfo.specifics}
@@ -170,6 +174,18 @@ export const SchoolTransportDetails: React.FC<SchoolTransportDetailsProps> = ({
                 />
               </View>
             )}
+            {directionsInfo.schoolName === 'Washington Elementary School' && (
+              <Image
+                source={require('../../assets/washington-1.png')}
+                style={styles.washingtonImage}
+              />
+            )}
+            {directionsInfo.schoolName === 'Washington Elementary School' && (
+              <Image
+                source={require('../../assets/washington-2.png')}
+                style={styles.washingtonImage}
+              />
+            )}
           </>
         )}
       </ScrollView>
@@ -194,14 +210,31 @@ const styles = StyleSheet.create({
     backgroundColor: 'white',
     alignItems: 'center',
     justifyContent: 'flex-start',
-    padding: 30,
+    paddingHorizontal: 30,
+    paddingBottom: 30,
     minHeight: screenHeight,
     flexGrow: 1,
     position: 'relative',
   },
-  section: {
-    margin: 0,
+  firstSection: {
+    paddingTop: 100,
     width: '100%',
+  },
+  section: {
+    width: '100%',
+  },
+  articleHeader: {
+    height: 350,
+    width: ScreenWidth,
+  },
+  overlay: {
+    position: 'absolute',
+    alignItems: 'center',
+    marginBottom: 150,
+    top: 300,
+    left: 20,
+    right: 20,
+    zIndex: 1,
   },
   header: {
     fontSize: 20,
@@ -213,11 +246,11 @@ const styles = StyleSheet.create({
     borderRadius: 20,
     padding: 20,
     width: '95%',
-    backgroundColor: 'rgba(255,228,181, 0.2)',
+    backgroundColor: 'rgba(255,228,181, 0.6)',
     marginBottom: 25,
   },
   headerTitle: {
-    fontSize: 26,
+    fontSize: 24,
     fontWeight: 'bold',
     marginBottom: 15,
   },
@@ -242,5 +275,10 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     borderRadius: 25,
     backgroundColor: 'black',
+  },
+  washingtonImage: {
+    width: 350,
+    height: 275,
+    marginTop: 20,
   },
 });
